PROYECTO_YOLOV8_SEGUNDO_PARCIAL/requirements.txt
# Dependencias del Proyecto de Detección de Objetos

# Framework de Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# YOLOv8 y Ultralytics
ultralytics>=8.0.0

# Procesamiento de imágenes y visión computacional
opencv-python>=4.8.0
Pillow>=9.5.0

# Análisis de datos y visualización
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Métricas y evaluación
scikit-learn>=1.3.0

# Jupyter notebooks
jupyter>=1.0.0
ipython>=8.0.0

# Utilidades
tqdm>=4.65.0
PyYAML>=6.0

# Manejo de archivos y datasets
requests>=2.31.0

# Opcional: Aceleración con GPU
# torch-audio  # Solo si se necesita procesamiento de audio
# tensorboard  # Para visualización de entrenamiento
